changes after review

+++ src/autopas/selectors/tuningStrategy/TuningStrategyFactory.cpp
@@ -19,9 +19,8 @@ std::unique_ptr<autopas::TuningStrategyInterface> autopas::TuningStrategyFactory
     const std::set<autopas::LoadEstimatorOption> &allowedLoadEstimators,
     const std::set<autopas::DataLayoutOption> &allowedDataLayouts,
     const std::set<autopas::Newton3Option> &allowedNewton3Options, unsigned int maxEvidence, double relativeOptimum,
-    unsigned int maxTuningPhasesWithoutTest, unsigned int relativeRangeForBlacklist,
-    unsigned int evidenceFirstPrediction, AcquisitionFunctionOption acquisitionFunctionOption,
-    ExtrapolationMethodOption extrapolationMethodOption) {
+    unsigned int maxTuningPhasesWithoutTest, double relativeRangeForBlacklist, unsigned int evidenceFirstPrediction,
+    AcquisitionFunctionOption acquisitionFunctionOption, ExtrapolationMethodOption extrapolationMethodOption) {
   // clang compiler bug requires static cast
   switch (static_cast<TuningStrategyOption>(tuningStrategyOption)) {
     case TuningStrategyOption::randomSearch: {
